image-metadata-api/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.5
Pillow>=9.0.0
python-magic-bin>=0.4.14